Move draft actions to single test file
