+++ codecept.conf.js
@@ -10,7 +10,7 @@ exports.config = {
   helpers: {
     Puppeteer: {
       show: process.env.SHOW_BROWSER_WINDOW || false,
-      restart: true,
+      restart: false,
       waitForTimeout: 5000,
       chrome: {
         ignoreHTTPSErrors: true,
