+++ e2e/config.js
@@ -36,19 +36,19 @@ module.exports = {
   },
   // actions
   applicationActions: {
-    selectOrders: 'Orders and directions needed',
-    selectHearing: 'Hearing needed',
+    enterOrdersAndDirectionsNeeded: 'Orders and directions needed',
+    enterHearingNeeded: 'Hearing needed',
     enterChildren: 'Children',
     enterRespondents: 'Respondents',
-    enterApplicants: 'Applicant',
+    enterApplicant: 'Applicant',
     enterOthers: 'Others to be given notice',
     enterGrounds: 'Grounds for the application',
-    enterRisk: 'Risk and harm to children',
+    enterRiskAndHarmToChildren: 'Risk and harm to children',
     enterFactorsAffectingParenting: 'Factors affecting parenting',
     enterInternationalElement: 'International element',
     enterOtherProceedings: 'Other proceedings',
     enterAllocationProposal: 'Allocation proposal',
-    attendingHearing: 'Attending the hearing',
+    enterAttendingHearing: 'Attending the hearing',
     uploadDocuments: 'Documents',
     changeCaseName: 'Change case name',
     submitCase: 'Submit application',
