+++ e2e/fixtures/children.js
@@ -1,0 +1,27 @@
+module.exports = [
+  {
+    address: {
+      lookupOption: 'Flat 2, Caversham House 15-17, Church Road, Reading',
+      buildingAndStreet: {
+        lineOne: 'Flat 2',
+        lineTwo: 'Caversham House 15-17',
+        lineThree: 'Church Road',
+      },
+      town: 'Reading',
+      postcode: 'RG4 7AA',
+      country: 'United Kingdom',
+    },
+  }, {
+    address: {
+      lookupOption: '2 Three Tuns Wynd, High Street, Stokesley, Middlesbrough',
+      buildingAndStreet: {
+        lineOne: '2 Three Tuns Wynd',
+        lineTwo: 'High Street',
+        lineThree: 'Stokesley',
+      },
+      town: 'Middlesbrough',
+      postcode: 'TS9 5DQ',
+      country: 'United Kingdom',
+    },
+  },
+];
