+++ e2e/fragments/addressPostcodeLookup.js
@@ -23,7 +23,9 @@ module.exports = {
     I.click(this.findAddressButton);
     I.waitForElement(locate(this.fields.addressList).find('option').withText(address.lookupOption));
     I.click(this.fields.addressList);
+    I.wait(1);
     I.selectOption(this.fields.addressList, address.lookupOption);
+    I.wait(1);
   },
 
   enterAddressManually(address) {
