+++ e2e/tests/old/caseSubmitted_test.js
@@ -1,4 +1,4 @@
-const config = require('../config.js');
+const config = require('../../config.js');
 const fields = {
   documentLink: 'ccd-read-document-field>a',
 };
