+++ e2e/tests/old/deleteApplication_test.js
@@ -1,4 +1,4 @@
-const config = require('../config.js');
+const config = require('../../config.js');
 
 Feature('Delete Case').retry(2);
