+++ e2e/tests/old/sendToGatekeeper_test.js
@@ -1,4 +1,4 @@
-const config = require('../config.js');
+const config = require('../../config.js');
 
 let caseId;
