+++ e2e/tests/old/uploadDocuments_test.js
@@ -1,39 +1,10 @@
-const config = require('../config.js');
+const config = require('../../config.js');
 
 Feature('Upload Documents');
 
 Before((I, caseViewPage) => {
   I.logInAndCreateCase(config.swanseaLocalAuthorityEmailUserOne, config.localAuthorityPassword);
   caseViewPage.goToNewActions(config.applicationActions.uploadDocuments);
-});
-
-Scenario('Selecting social work chronology document to follow in the c110a application', (I, uploadDocumentsEventPage, caseViewPage) => {
-  uploadDocumentsEventPage.selectSocialWorkChronologyToFollow();
-  I.continueAndSave();
-  I.seeEventSubmissionConfirmation(config.applicationActions.uploadDocuments);
-  caseViewPage.selectTab(caseViewPage.tabs.documents);
-  I.seeDocument('Social work chronology', '', 'To follow', 'mock reason');
-});
-
-Scenario('Uploading all files in the c110a application', (I, uploadDocumentsEventPage, caseViewPage) => {
-  uploadDocumentsEventPage.selectSocialWorkChronologyToFollow(config.testFile);
-  uploadDocumentsEventPage.uploadSocialWorkStatement(config.testFile);
-  uploadDocumentsEventPage.uploadSocialWorkAssessment(config.testFile);
-  uploadDocumentsEventPage.uploadCarePlan(config.testFile);
-  uploadDocumentsEventPage.uploadSWET(config.testFile);
-  uploadDocumentsEventPage.uploadThresholdDocument(config.testFile);
-  uploadDocumentsEventPage.uploadChecklistDocument(config.testFile);
-  uploadDocumentsEventPage.uploadAdditionalDocuments(config.testFile);
-  I.continueAndSave();
-  I.seeEventSubmissionConfirmation(config.applicationActions.uploadDocuments);
-  caseViewPage.selectTab(caseViewPage.tabs.documents);
-  I.seeDocument('Social work chronology', '', 'To follow', 'mock reason');
-  I.seeDocument('Social work statement and genogram', 'mockFile.txt', 'Attached');
-  I.seeDocument('Social work assessment', 'mockFile.txt', 'Attached');
-  I.seeDocument('Care plan', 'mockFile.txt', 'Attached');
-  I.seeDocument('Social work evidence template (SWET)', 'mockFile.txt', 'Attached');
-  I.seeDocument('Threshold document', 'mockFile.txt', 'Attached');
-  I.seeDocument('Checklist document', 'mockFile.txt', 'Attached');
 });
 
 Scenario('As a local authority I have the ability to upload a document after submission of a case', (I, uploadDocumentsEventPage, submitApplicationEventPage, caseViewPage) => {
